requirements.txt
django
pip install django-jazzmin==2.6.1
django-jazzmin
django-admin-charts
python-dotenv